+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Master
 
+- Adds new base container / screen for the Collection page - luc
 - Fixes intermittent crash on opening message attachment previews - ash
 - Fix consignments artist search and make it use relay - alloy
 - Adds links to Artwork view for FAQ and Ask a Specialist - kierangillen